% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read}
\alias{read}
\title{Reads a data set from jamovi}
\usage{
read(id, columns)
}
\arguments{
\item{id}{the number, or the title of the data set to read}

\item{columns}{(optional) only reads the columns named}
}
\value{
the data set as a data frame
}
\description{
Reads a data set from jamovi
}
\examples{
\dontrun{
jmvconnect::what()

#  Available Data Sets
#  -------------------------------------
#         Title           Rows    Cols
#  -------------------------------------
#    1    iris             150       5
#    2    Tooth Growth      60       3
#  -------------------------------------

data <- jmvconnect::read('Tooth Growth')

# or

data <- jmvconnect::read(2)
}
}
