% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{what}
\alias{what}
\title{Lists the data sets available from jamovi}
\usage{
what()
}
\description{
Lists the data sets available from jamovi. Data sets can then be read using
the read() function.
}
\examples{
\dontrun{
jmvconnect::what()

#  Available Data Sets
#  -------------------------------------
#         Title           Rows    Cols
#  -------------------------------------
#    1    iris             150       5
#    2    Tooth Growth      60       3
#  -------------------------------------

data <- jmvconnect::read('Tooth Growth')

# or

data <- jmvconnect::read(2)
}

}
